# docs/swagger.yaml
openapi: "3.0.3"
info:
  title: "Asistencia SaaS API"
  description: "API REST para control de asistencia y gestión de personal multi-tenant"
  version: "1.0.0"
  contact:
    name: "Asistencia SaaS"
    url: "https://asistencia-saas.com"

servers:
  - url: http://localhost:3001/api/v1
    description: Desarrollo local
  - url: https://api.asistencia-saas.com/api/v1
    description: Producción

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token JWT obtenido del login"

  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
        timestamp:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object

    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20

    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    RegisterDto:
      type: object
      required:
        - companyName
        - adminEmail
        - adminPassword
        - adminFirstName
        - adminLastName
      properties:
        companyName:
          type: string
          maxLength: 255
        legalName:
          type: string
        rut:
          type: string
        adminEmail:
          type: string
          format: email
        adminPassword:
          type: string
          minLength: 8
        adminFirstName:
          type: string
        adminLastName:
          type: string
        phone:
          type: string
        timezone:
          type: string
          default: "America/Lima"

    CheckInDto:
      type: object
      required:
        - employeeId
        - latitude
        - longitude
      properties:
        employeeId:
          type: string
          format: uuid
        latitude:
          type: number
          format: float
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: float
          minimum: -180
          maximum: 180
        accuracy:
          type: number
          description: "Precisión GPS en metros"
        method:
          type: string
          enum: [GPS, MANUAL, QR]
          default: GPS
        photo:
          type: string
          format: byte
          description: "Foto en base64"
        projectId:
          type: string
          format: uuid
        deviceInfo:
          type: object

    CheckOutDto:
      type: object
      required:
        - employeeId
        - latitude
        - longitude
      properties:
        employeeId:
          type: string
          format: uuid
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        accuracy:
          type: number
        method:
          type: string
          enum: [GPS, MANUAL, QR]
          default: GPS
        photo:
          type: string
          format: byte
        notes:
          type: string

    CreateEmployeeDto:
      type: object
      required:
        - firstName
        - lastName
        - documentType
        - documentNumber
      properties:
        code:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        documentType:
          type: string
          enum: [DNI, CE, Pasaporte]
        documentNumber:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        position:
          type: string
        department:
          type: string
        hireDate:
          type: string
          format: date
        workScheduleId:
          type: string
          format: uuid
        emergencyContact:
          type: string
        emergencyPhone:
          type: string

  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
    DateFromParam:
      name: dateFrom
      in: query
      schema:
        type: string
        format: date
    DateToParam:
      name: dateTo
      in: query
      schema:
        type: string
        format: date

paths:
  /health:
    get:
      tags: [Health]
      summary: "Health check del servidor"
      responses:
        "200":
          description: "Servidor saludable"

  /auth/login:
    post:
      tags: [Auth]
      summary: "Iniciar sesión"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: "Login exitoso"
        "401":
          description: "Credenciales inválidas"

  /auth/register:
    post:
      tags: [Auth]
      summary: "Registrar nueva empresa"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        "201":
          description: "Registro exitoso"

  /auth/refresh:
    post:
      tags: [Auth]
      summary: "Renovar access token"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: "Token renovado"

  /attendance/check-in:
    post:
      tags: [Attendance]
      summary: "Registrar entrada"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInDto"
      responses:
        "201":
          description: "Check-in registrado"
        "400":
          description: "Error de validación (fuera de geocerca, etc)"

  /attendance/check-out:
    post:
      tags: [Attendance]
      summary: "Registrar salida"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutDto"
      responses:
        "200":
          description: "Check-out registrado"

  /attendance:
    get:
      tags: [Attendance]
      summary: "Historial de asistencia"
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [ON_TIME, LATE, MISSING]
      responses:
        "200":
          description: "Lista de asistencias"

  /employees:
    get:
      tags: [Employees]
      summary: "Listar empleados"
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, SUSPENDED]
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: "Lista de empleados"

    post:
      tags: [Employees]
      summary: "Crear empleado"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeDto"
      responses:
        "201":
          description: "Empleado creado"

  /geolocation/fences:
    get:
      tags: [Geolocation]
      summary: "Listar cercas virtuales"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: "Lista de geocercas"

    post:
      tags: [Geolocation]
      summary: "Crear cerca virtual"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, latitude, longitude, radius]
              properties:
                name:
                  type: string
                address:
                  type: string
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
                radius:
                  type: number
                  description: "Radio en metros"
                workScheduleId:
                  type: string
                  format: uuid
      responses:
        "201":
          description: "Geocerca creada"

  /geolocation/verify:
    post:
      tags: [Geolocation]
      summary: "Verificar ubicación en geocerca"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
                fenceId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: "Resultado de verificación"

  /overtime/request:
    post:
      tags: [Overtime]
      summary: "Solicitar horas extra"
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [employeeId, date, startTime, endTime, reason]
              properties:
                employeeId:
                  type: string
                  format: uuid
                date:
                  type: string
                  format: date
                startTime:
                  type: string
                  example: "18:00"
                endTime:
                  type: string
                  example: "20:30"
                reason:
                  type: string
      responses:
        "201":
          description: "Solicitud creada"